The final product is in sight. It is possible to add various movies to the "queue" in order to compare them, the api for the actors of a movie is implemented.
The only things which I need to do are removing movies styling the movies a bit better (using the 'cards' styleguide perhaps), and joining the artists.

Also, a couple of other things, which I do not know yet.

What could happen later is using various sources, i.e. IMDB and TMDB and FreeBase. For now it is just the TMDB. :D