I want to start using the rovi api, because that seems extensible enough. :D (All those API's, I really need to settle on one. >:D).
Anyhow, I'm happy with the movie cards, but not yet so happy with the cast cards. I really need to come up with a nice way to present the cast of a movie.

Actually, come to think of it, I don't really need that, because I just need to present the cast playing in the selected movies. This will be later though. :D